refactor: 모달 파일명이 일관되지 않던 문제점 해결
